refactor(error): remove unused variants, move Error to its own mod
