+++ Cargo.toml
@@ -23,7 +23,6 @@ tokio = { version = "1.0", features = ["macros", "sync", "time", "rt", "rt-multi
 rand = "0.7"
 rusqlite = { version = "0.24", features = ["bundled"], optional = true }
 slip10 = "0.4"
-bech32 = "0.7"
 hex = "0.4"
 futures = "0.3"
 backtrace = "0.3"
